Finished FBX + Textures + Animations
It took me a while to get out of the burnout, but at least I finished an idle, walking and attack animations for both the robot and the playable character.
All textures, normals and Uv's should work just be, and I separated the animations in Actions inside of blender to make the implementation easier while on Unity.

(I'll get to it in a separate Unity project to check a couple things, after that I'll probably ask a couple things and apply it to the main file tomorrow).

[I wasn't sure if I should replace the Rusty_Robot model or not, so I just added the updated version besides the other model, just tell me and I'll delete the old one and rename the new one, no problem].

{Yes, I know there is a material folder, but I think I'll need a separate one to put the textures into, even if only temporary (I think, if not, where should I put it?)}.