Added new animations
I updated the Idle and run animation, as well as adding a melee and shoot attack animations. I may have to re-do the shooting one tho; tell me what you think of it.
Also, I was unsure on how to do the dash animation, so I added a "Test" one that's practically the other one I added but inverted. Hope that works.